Move more sampling logic in SampleSet.
A lot of this might just be shifting stuff from animated_mesh_ctrl to SampleSet,
but it seems to help separate sampling from Animesh.  The main interaction points
now are Animated_mesh_ctrl::update_bone_samples, where samples are loaded
into Animesh, and HRBF_sampling, which reads vertex data from Animesh.  The
rest is either self-contained or only needs a Skeleton.

(A lot of this is just an exercise to help me understand the data flow.)