Reduce the Bone hierarchy to one class.
This makes Bone, Bone_hrbf, Bone_precomputed and Bone_ssd a single class with
a couple state flags ("hrbf on/off", "precomputed yes/no").

This keeps the non-precomputed HRBF data around, and the precomputed data is
just a cache that can be toggled on and off, without recreating the bone itself.

This also exposes a non-const Skeleton.get_bone.  Although there's a loud comment
saying changes to bones need to go through set_bone because it updates things
based on the changes, set_bone doesn't actually do that, it just stores the new bone.
There's a commented-out "to be deleted update_hrbf_id_to_bone_id()" call in
there, so this looks like something that used to be the case but was factored
out.