Refactor the way bones are created from joints.
Previously, each joint had a length derived from the average direction of the
joint to each of its children.  The position of the joint down the direction's
length was the bone.  This wasn't really correct: if a joint had two children
going in exactly opposite directions, the average direction would be zero, giving
a zero orientation and leading to division by zero later.  Having exactly symmetrical
children isn't unlikely with a model in T-pose.

Instead, treat each bone as starting from its parent's joint position and going
to the joint position.  There's no averaging child directions.  This matches how bones
are structured in Maya.  If a joint has two symmetrical children, you have two
separate symmetrical bones.  Bones are transformed by the parent joint's transform.

This means that rather than skipping leaves (is_leaf) when working with bones,
we need to skip the first joint.  Before the leaf joints didn't have a bone
associated with them, but now the first joint is the one with no bone.  (The
root joint also doesn't create a bone.)  Use Skeleton.has_bone to test this.

This also fixes errors if a joint lies on the same position as its parent, causing
a zero orientation and IND's.