Refactor Precomputed_prim using texture objects (experimental).
Use texture objects so we can encapsulate Precomputed_prim, which allows it to
be done more simply.  We have a separate grid for each one, instead of copying
them all into one big grid.  This means offsets go away, and we can use texture
clamping to avoid the need to check the boundaries.

This is experimental because it's probably slower, and it requires SM3.0, but
the maintainability win is pretty huge.  We could probably get that without
the performance hit by going back to textures for everything, but using texture
objects instead of texture references.