Refactor SampleSet.
This makes the joint cap settings a sample-time value, instead of something
that gets persisted.  They'll just be UI values when performing the sampling;
if you want to change them later, resample the whole bone.  This means that
any editing of the other samples would be lost, but if you're editing the
samples, you could just adjust the cap samples too.

The joint cap samples get stored like any other sample, rather than having
a separate bit of data just for the caps.

The goal of this is to make it easier to implement sample editing and to
reduce the types of data that need to be saved, and to make things line up
with the way I think the UI and data model will work.

This removes factor_bones.  It could come back later, but for now I'm reducing
this a bit to make implementation easier.