Refactor bone IDs.
Previously, bone IDs were the index of the bone in the skeleton.  This made bones
inherently skeleton-specific.

Now, bones are assigned a unique ID when they're created, which is independent
of any skeleton.  Skeleton, Tree, etc. now store bones as a map<id,Bone*>, and
Skeleton's ref to the Bone is always const.

This is meant to make it possible for bones to exist independently of skeletons,
or to be used with more than one skeleton at a time.  It's not quite there yet,
but that's now external to the deformer logic.

Many operations will still require a Skeleton even though they'll work on a
single Bone (eg. sampling), but we can just create a temporary skeleton.