Work on making Bones work outside of a full skeleton.
This sets up a small Skeleton inside each Bone, which is only used to
be able to work with each bone independently.  For example, Bone->precompute
doesn't need to be given a skeleton since it's not actually skeleton-
specific.  It just uses its own, since the code is designed around
it.

Places that we modify the Bone are moved into BoneSet; Animated_mesh_ctrl
is const on Bone, and is basically just a thin wrapper around Animesh
since we still can't load the Animesh header alongside Maya headers.

hd_bone_arrays->hd_bone_hrbf and hd_bone_arrays->hd_bone_precomputed
now store the IDs regardless of the state of the bone.  We always
have HRBF and Precomputed environments set up for all bones, even if
they're in a different state (it's simpler to set them up and leave
them set up).  This makes things like filling in the precompute
grid work even if we're already in the precompute state; previously
the HRBF ID would be -1 in this case.