Working on refactoring the plugin structure.
- Remove unskinnedGeomAttr, and always just load the input geometry.
Previously we needed to be able to load the unskinned geometry (the
input into the skinCluster) to calculate base potential, but we do
that on setup and store it now.  This reduces how tightly connected
we are to skinCluster.

- Add a hack to avoid reloading the mesh entirely whenever it changes.
This wasn't needed with the separate unskinnedGeomAttr connection.

- Load base potential during the mesh load, instead of as a separate
dependency.  The above hack doesn't prevent base potential from being
marked dirty, so we'd reload base potential on every frame.