Process everything in world space, instead of object space.
This is working towards separating the implicit surfaces from the deformer,
so they'll exist as independent objects.  Since they won't have a hard
relationship with the object being deformed, it doesn't make sense for them
to be in its object space.  Store them in world space.

This was in object space before since we weren't storing base potential,
so doing it in object space allowed us to get the unskinned geometry from
the skinCluster.