Major refactoring of the plugin layout.
This became too big a change to split apart, so this is a monolithic
commit.

- Surfaces, blending and deformers are split into separate nodes.  Surfaces
connect to blending nodes (or deformers, for debuggin), and blend nodes
connect to deformers.  Implicit surfaces can exist on their own, with
no deformation happening.

This fits better in Maya's model, and will make it easier to implement
editing and other UIs.

- Surfaces are positioned as regular Maya shapes.  We don't do anything
special now except to propagate the transforms into the hrbf/precomp.
- Bones and skeletons are stored in shared_ptr.  Blend nodes and deformers
need to hold a reference to a skeleton owned by the surface, and Maya can
sometimes delete objects out of order (eg. deleting a bone before deleting
a deformer that uses it).

This isn't done all the way down the chain; places where we grab a Bone*
locally can keep doing what they're doing.