Try to clarify match_base_potential a bit.
Scope the final r(t) to where it's needed.  Previously, if no binary search
was needed, r(t) would do the same thing as the last r(t) inside the loop
and give the same value, and if nb_iter was 0 then t==0 and it wouldn't do
anything.  The only case it's actually needed is to apply the result of
binary_search() (or the commented-out binary_search_div).  Make this clear
by doing the r(t) right after binary_search().  This allows scoping the
variables more tightly.

Move the zero gradient check to the start of the loop, which allows combining
the two; there was one outside the loop and another at the end of the loop.