Remove the final_pass parameter to match_base_potential.
All it does is prevent vert_to_fit from being updated when it's not
going to be used.  It's an extra conditional in the inner loop to
save a write in the outer loop, so it doesn't seem worth it.

From comments elsewhere, this flag looks like it used to do something
else ("First fitting (only evaluate the three adjacent nearest bones)").