Disable "warning C4267: 'initializing' : conversion from 'size_t' to 'int', possible loss of data".
This causes hundreds of warnings in x64 builds.  It's a bogus warning, since it
suggests that people should change int to size_t.  size_t is unsigned, so using
it blindly is a source of subtle bugs.  Unless you're storing pointers or really
need fast division by powers of 2, there's rarely a reason to use unsigned 64-bit
ints.  This is a warning from 32-bit days which is just a nuisance in 64-bit, so
get rid of it.

(This doesn't disable double-to-float warnings, which have a different code.)