Fix the precomputed grid coordinate space.
Precomputed_prim::fill_grid_with was reading the bone it's precomputing
in world space, which is incorrect; the precompute is always in object
space, since we don't want to recalculate it every time a bone moves.
This was being masked by the fact that it was being computed before
the world space matrices were actually set, but the problems became
visible if an implicit surface's sample points were modified later.