Fix loading base potential not fully using the correct data path.
Previously when we updated base potential, we'd update it on the GPU
data in-place, then read it out and save it to the node.  Now, we
update it to a new buffer, save it to the node, then the dirty node
causes the new base potential to be loaded into the GPU.  This way,
base potential is always loaded in the same way.

This also gets rid of the setInternalValueInContext hack, using
setDependentsDirty instead.

Fixes changes to base potential in the AE not taking effect.