Fix the skeleton grids being updated too much.
When a surface is moved, we need to update the skeletons pointing at
it.  This was happening on each world matrix load, so the skeleton
could be updated 20 times if you have 20 surfaces.  Fix this by putting
a sequence number on each bone which is incremented when a change is
made that requires bones to update, so we can update the skeleton just
once when it's actually used.