update_device_grid optimizations.
The grid has 20^3=8000 cells (the number of filled cells is smaller).  However,
we only actually have a very small number of unique blending lists, relative
to the number of surfaces.  Some time is being spent recreating that list for
each cell.  Optimize this a bit by first calculating all of the blending lists
for each cluster, then cell_to_blending_list just creates a list of pointers
to those lists, instead of calculating the list and then appending the list to
blist.

This optimization is a bit ugly, but brings my current test case from ~12 FPS
to ~13 FPS, which is a decent percentage.