Flush the CUDA dispatch queue at a strategic point to improve parallelism.
This probably only makes much of a difference on Windows, where the kernel
dispatch overhead is much higher than on Linux.  This pushes the fit_mesh
call to the GPU right away, since it usually takes a relatively long time,
so the rest of the dispatches can happen in parallel with it.  On later
passes (where there aren't many vertices to fit and the fit_mesh call takes
less time) the dispatch overhead is still a factor, but this helps a bit.

Also, perform the readback last, instead of before pack<>.