Make it easier to enable error checking.
Enabling error checking slows things down a lot, since it prevents pipelining,
but currently it can only be enabled in a debug build, which is even slower.
Most development happens in release builds, so this makes it easy to miss
errors.  Move the error checking logic into a source file instead of the
header, so it can be turned on and off independently and without rebuilding
the whole project.

Also, enable CUDA_SAFE_CALL all the time.  That doesn't require a cudaThreadSynchronize,
so there's no performance hit.